+++ packages/music-module/.prettierrc.cjs
@@ -3,7 +3,7 @@
 module.exports = require("@leomotors/config/prettier").withSortImports({
   importOrder: [
     "^dotenv",
-    "^cocoa-discord-utils",
+    "^cocoa-discord",
     "^discord.js",
     "^@discordjs",
     "^[a-zA-Z]",