+++ packages/music-module/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./config.g.js";
 
-export { CocoaVersion } from "cocoa-discord-utils/meta";
+export { CocoaVersion } from "cocoa-discord/meta";
 
 export { Music } from "./music.resolver.js";
 export * as Service from "./music.service.js";